handle checkbox added to functionality
